Convert Input Plugin to Blueprint Library
Since there is little documentation about how an Input
Plugin works with Unreal, I was having a hard time
figuring it out. I had a stroke of inspiration for a
different way to handle input. It kind of circumvents
Unreal's input methodology, but I think it will work
just fine for what we're trying to do.